labfiles/foundry-chat/python/requirements.txt
python-dotenv
openai
